+++ NEWS.md
@@ -1,3 +1,21 @@
+## version 0.5.1
+
+---
+
+- fix format for initial NEWS.md line
+
+## version 0.5.0
+
+---
+
+- add function to combine dev version's points
+
+## version 0.4.1
+
+---
+
+- fix testing setup
+
 ## version 0.4.0
 
 ---
@@ -68,5 +86,5 @@
 
 ### NEWS.md setup
 
-- added NEWS.md creation with newsmd
+- added NEWS.md creation with [newsmd](https://github.com/Dschaykib/newsmd)